style(subway): separate each class
- 하나의 클래스를 다른 하나의 클래스로 분리
