+++ src/main/java/subway/controller/SubwayController.java
@@ -3,7 +3,9 @@ package subway.controller;
 import subway.service.LineService;
 import subway.service.StationService;
 import subway.service.SubwayService;
-import subway.service.TransitMapService;
+import subway.service.initialization.LineInitialization;
+import subway.service.initialization.StationInitialization;
+import subway.service.initialization.TransitMapInitialization;
 
 import java.util.Scanner;
 
@@ -14,9 +16,9 @@ public class SubwayController {
     }
 
     public static void initializeSubway() {
-        StationService.initializeStations();
-        LineService.initializeLines();
-        TransitMapService.initializeTransitMaps();
+        StationInitialization.initializeStations();
+        LineInitialization.initializeLines();
+        TransitMapInitialization.initializeTransitMaps();
     }
 
     public static void startSubway(Scanner scanner) {
